Fix request addition and time formatting

+++ app/src/main/java/keronei/swapper/data/domain/BatteryRequestUpdateModel.kt
@@ -1,10 +1,10 @@
 package keronei.swapper.data.domain
 
 class BatteryRequestUpdateModel(
-    val id: Int?,
+    val id: Long?,
     val batteryCount: Int,
     val batteries: List<String>,
-    val requestId: Int,
+    val requestId: Long,
     val comment: String,
     val time: String
 )
