+++ app/src/main/java/keronei/swapper/data/local/BatteryRequestDao.kt
@@ -8,13 +8,13 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface BatteryRequestDao {
     @Insert
-    suspend fun createBatteryRequest(batteryRequest: BatteryRequestsEntity): Int
+    suspend fun createBatteryRequest(batteryRequest: BatteryRequestsEntity): Long
 
     @Insert
-    suspend fun updateBatteryRequest(batteryRequestUpdate: BatteryRequestUpdateEntity): Int
+    suspend fun updateBatteryRequest(batteryRequestUpdate: BatteryRequestUpdateEntity): Long
 
     @Query("DELETE from batteryrequestsentity WHERE id = :batteryRequestId")
-    suspend fun removeBatteryRequest(batteryRequestId: Int): Int
+    suspend fun removeBatteryRequest(batteryRequestId: Long): Int
 
     @Query("SELECT * from batteryrequestsentity")
     fun getLocalRequests(): Flow<List<UpdatedRequestsEmbed>>
